requirements.txt
ansible==2.3.1.0
asn1crypto==0.22.0
awscli==1.11.127
bcrypt==3.1.3
boto==2.48.0
boto3==1.4.4
botocore==1.5.90
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
colorama==0.3.7
cryptography==2.0.2
Django==1.10.4
djangorestframework==3.5.3
docutils==0.13.1
elasticsearch==5.0.1
enum34==1.1.6
fix-yahoo-finance==0.0.18
futures==3.1.1
idna==2.5
ipaddress==1.0.18
Jinja2==2.9.6
jmespath==0.9.3
MarkupSafe==1.0
multitasking==0.0.4
numpy==1.11.3
pandas==0.20.3
pandas-datareader==0.5.0
paramiko==2.2.1
pyasn1==0.3.1
pycparser==2.18
pycrypto==2.6.1
PyNaCl==1.1.2
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.4
requests-file==1.4.2
requests-ftp==0.3.1
rsa==3.4.2
s3transfer==0.1.10
six==1.10.0
slugify==0.0.1
urllib3==1.22
uWSGI==2.0.15
